+++ Exposure/Authentication/Response/SessionToken.swift
@@ -98,8 +98,9 @@ extension SessionToken {
     }
     
     /// Returns true if the session token has the correct format
+    @available(*, deprecated: 2.0.89, message: "Token format is considered valid for all strings.")
     public var hasValidFormat: Bool {
-        return components.count == 9
+        return true
     }
 }